Crud PreEjecucionPresupuestaria

+++ Convertidor/Data/Entities/Presupuesto/PRE_EJECUCION_PRESUPUESTARIA.cs
@@ -8,14 +8,14 @@
         public string CODIGO_NIVEL2 { get; set; } = string.Empty;
         public string CODIGO_NIVEL3 { get; set; } = string.Empty;
         public string CODIGO_NIVEL4 { get; set; } = string.Empty;
-        public int I_REAL { get; set; }
-        public int I_PROYECTADO { get; set; }
-        public int II_REAL { get; set; }
-        public int II_PROYECTADO { get; set; }
-        public int III_REAL { get; set; }
-        public int III_PROYECTADO { get; set; }
-        public int IV_REAL { get; set; }
-        public int IV_PROYECTADO { get; set; }
+        public decimal I_REAL { get; set; }
+        public decimal I_PROYECTADO { get; set; }
+        public decimal II_REAL { get; set; }
+        public decimal II_PROYECTADO { get; set; }
+        public decimal III_REAL { get; set; }
+        public decimal III_PROYECTADO { get; set; }
+        public decimal IV_REAL { get; set; }
+        public decimal IV_PROYECTADO { get; set; }
         public string EXTRA1 { get; set; } = string.Empty;
         public string EXTRA2 { get; set; } = string.Empty;
         public string EXTRA3 { get; set; } = string.Empty;
